feat: Refactor audio download functionality and replace Download context with AudioDownload context
